refactor: extract visualization logic from model
- Create `src/mlcast/visualization.py` to house visualization utilities.
- Move `apply_radar_colormap` and `log_images` out of `convgru.py`.
- Update `RadarLightningModel` to use the extracted `log_images`.
